Convert floats that interact with the UI to CGFloats. (#284)
* Convert floats that interact with the UI to CGFloats.

+++ Additions/NSAttributedStringAdditions.m
@@ -415,14 +415,14 @@ NSString *NSChatCTCPTwoFormatType = @"NSChatCTCPTwoFormatType";
 				if( scanOneOrTwoDigits( scanner, &fcolor ) ) {
 					fcolor %= 16;
 
-					NSColor *foregroundColor = [NSColor colorWithCalibratedRed:( (float) mIRCColors[fcolor][0] / 255. ) green:( (float) mIRCColors[fcolor][1] / 255. ) blue:( (float) mIRCColors[fcolor][2] / 255. ) alpha:1.];
+					NSColor *foregroundColor = [NSColor colorWithCalibratedRed:( (CGFloat) mIRCColors[fcolor][0] / 255. ) green:( (CGFloat) mIRCColors[fcolor][1] / 255. ) blue:( (CGFloat) mIRCColors[fcolor][2] / 255. ) alpha:1.];
 					if( foregroundColor && ! [[options objectForKey:@"IgnoreFontColors"] boolValue] )
 						[attributes setObject:foregroundColor forKey:NSForegroundColorAttributeName];
 
 					NSUInteger bcolor = 0;
 					if( [scanner scanString:@"," intoString:NULL] && scanOneOrTwoDigits( scanner, &bcolor ) && bcolor != 99 ) {
 						bcolor %= 16;
-						NSColor *backgroundColor = [NSColor colorWithCalibratedRed:( (float) mIRCColors[bcolor][0] / 255. ) green:( (float) mIRCColors[bcolor][1] / 255. ) blue:( (float) mIRCColors[bcolor][2] / 255. ) alpha:1.];
+						NSColor *backgroundColor = [NSColor colorWithCalibratedRed:( (CGFloat) mIRCColors[bcolor][0] / 255. ) green:( (CGFloat) mIRCColors[bcolor][1] / 255. ) blue:( (CGFloat) mIRCColors[bcolor][2] / 255. ) alpha:1.];
 						if( backgroundColor && ! [[options objectForKey:@"IgnoreFontColors"] boolValue] )
 							[attributes setObject:backgroundColor forKey:NSBackgroundColorAttributeName];
 					}
@@ -525,7 +525,7 @@ NSString *NSChatCTCPTwoFormatType = @"NSChatCTCPTwoFormatType";
 							NSUInteger index = [colorStr characterAtIndex:0];
 							if( index >= 'A' ) index -= ( 'A' - '9' - 1 );
 							index -= '0';
-							NSColor *foregroundColor = [NSColor colorWithCalibratedRed:( (float) CTCPColors[index][0] / 255. ) green:( (float) CTCPColors[index][1] / 255. ) blue:( (float) CTCPColors[index][2] / 255. ) alpha:1.];
+							NSColor *foregroundColor = [NSColor colorWithCalibratedRed:( (CGFloat) CTCPColors[index][0] / 255. ) green:( (CGFloat) CTCPColors[index][1] / 255. ) blue:( (CGFloat) CTCPColors[index][2] / 255. ) alpha:1.];
 							if( foregroundColor && ! [[options objectForKey:@"IgnoreFontColors"] boolValue] )
 								[attributes setObject:foregroundColor forKey:NSForegroundColorAttributeName];
 						} else if( [scanner scanString:@"." intoString:NULL] ) { // reset the foreground color
@@ -550,7 +550,7 @@ NSString *NSChatCTCPTwoFormatType = @"NSChatCTCPTwoFormatType";
 								NSUInteger index = [colorStr characterAtIndex:0];
 								if( index >= 'A' ) index -= ( 'A' - '9' - 1 );
 								index -= '0';
-								NSColor *backgroundColor = [NSColor colorWithCalibratedRed:( (float) CTCPColors[index][0] / 255. ) green:( (float) CTCPColors[index][1] / 255. ) blue:( (float) CTCPColors[index][2] / 255. ) alpha:1.];
+								NSColor *backgroundColor = [NSColor colorWithCalibratedRed:( (CGFloat) CTCPColors[index][0] / 255. ) green:( (CGFloat) CTCPColors[index][1] / 255. ) blue:( (CGFloat) CTCPColors[index][2] / 255. ) alpha:1.];
 								if( backgroundColor && ! [[options objectForKey:@"IgnoreFontColors"] boolValue] )
 									[attributes setObject:backgroundColor forKey:NSBackgroundColorAttributeName];
 							} else if( [scanner scanString:@"." intoString:NULL] ) { // reset the background color
